+++ spfw/log.h
@@ -1,7 +1,7 @@
 #ifndef __LOG__
 #define __LOG__
 
-void logToSerial(long load_cell_reading, int position, double speed);
+void logToSerial(long load_cell_reading, unsigned long position, double speed);
 
 #endif
